+++ cli/tests/test_config_command.rs
@@ -1252,26 +1252,16 @@ fn test_config_get() {
     ");
 
     let output = test_env.run_jj_in(".", ["config", "get", "table.list"]);
-    insta::assert_snapshot!(output, @r"
-    ------- stderr -------
-    Config error: Invalid type or value for table.list
-    Caused by: Expected a value convertible to a string, but is an array
-    Hint: Check the config file: $TEST_ENV/config/config0002.toml
-    For help, see https://jj-vcs.github.io/jj/latest/config/ or use `jj help -k config`.
+    insta::assert_snapshot!(output, @r#"
+    ["list", "value"]
     [EOF]
-    [exit status: 1]
-    ");
+    "#);
 
     let output = test_env.run_jj_in(".", ["config", "get", "table"]);
-    insta::assert_snapshot!(output, @r"
-    ------- stderr -------
-    Config error: Invalid type or value for table
-    Caused by: Expected a value convertible to a string, but is a table
-    Hint: Check the config file: $TEST_ENV/config/config0003.toml
-    For help, see https://jj-vcs.github.io/jj/latest/config/ or use `jj help -k config`.
+    insta::assert_snapshot!(output, @r#"
+    { string = "some value 1", int = 123, list = ["list", "value"], overridden = "bar" }
     [EOF]
-    [exit status: 1]
-    ");
+    "#);
 
     let output = test_env.run_jj_in(".", ["config", "get", "table.overridden"]);
     insta::assert_snapshot!(output, @r"